% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtc_colors.R
\name{color_jtc}
\alias{color_jtc}
\title{Add PYN colors to a chart}
\usage{
color_jtc(type = "fill", scale = "jtc", rev = FALSE)
}
\arguments{
\item{type}{"fill" for use with the fill aesthetic; 
"color" for use with the color aesthetic}

\item{scale}{"jtc" (default)}

\item{rev}{reverse the order of the colors; defaults to false}
}
\value{
a manual scale object
}
\description{
Theme has 8 colors. If you need more than that, you will need to set theme 
directly using scale_*_manual(values = jtc_pal[[XX]])
}
