% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyn_colors.R
\name{color_pyn}
\alias{color_pyn}
\title{Add PYN colors to a chart}
\usage{
color_pyn(type = "fill", scale = "pyn", rev = FALSE)
}
\arguments{
\item{type}{"fill" for use with the fill aesthetic; 
"color" for use with the color aesthetic}

\item{scale}{"pyn" (default), "reds", "yellows", "blues", "greys"}

\item{rev}{reverse the order of the colors; defaults to false}
}
\value{
a manual scale object
}
\description{
"PYN" theme has seven colors, the others have 4. If you need more than that,
you will need to set theme directly using 
scale_*_manual(values = pyn_pal[[XX]])
}
