% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{count_all}
\alias{count_all}
\title{Count
counts of variables for each non-numeric column}
\usage{
count_all(dat)
}
\value{
a list of counts for each non-numeric column
}
\description{
Count
counts of variables for each non-numeric column
}
