% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtc_colors.R
\name{create_pallete}
\alias{create_pallete}
\title{Create color pallete}
\usage{
create_pallete()
}
\value{
a list of collor vectors saved to the global environment
}
\description{
Adds color pallete to gloabl enviornment
already saved in the data, but this function allows 
pallete to be edited later
}
