% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_utils.R
\name{fix_names}
\alias{fix_names}
\title{Fix Names}
\usage{
fix_names(df)
}
\description{
Changes the column names of a data frame, removing any non-alphanumeric
characters and changing them to an underscore
}
