% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prev_month.R
\name{get_prev_month}
\alias{get_prev_month}
\title{Get Previous Month}
\usage{
get_prev_month(x)
}
\value{
a string containing the numeric value of the pervious month
}
\description{
gets the current month and subtracts one month b/c pulling data from
previous month
input doesn't matter, will always give previous month in a character string
with a leading 0 for single digit months
}
