% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{make_pct}
\alias{make_pct}
\title{Make percent}
\usage{
make_pct(x, dec = 1)
}
\arguments{
\item{dec}{number of decimal places to include; defaults to one}
}
\description{
converts real number to a string with a percent sign
good for formatting tables and charts for presentation
}
