% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_vars.R
\name{na_to}
\alias{na_to}
\title{NA to}
\usage{
na_to(var, rep)
}
\value{
a vector with NA values change to replacement value
}
\description{
A function that changes an NA variable to any replacement that is specified
}
