% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pyn.R
\name{theme_pyn}
\alias{theme_pyn}
\title{PYN Theme}
\usage{
theme_pyn(base_family = "Century Gothic", base_size = 14,
  axis.title.y = element_text(hjust = 1, vjust = 2),
  axis.title.x = element_text(hjust = 1), plot.title = element_text(hjust =
  0, size = 20), plot.subtitle = element_text(hjust = 0, size = 16),
  legend.position = "right", background = "blank")
}
\arguments{
\item{background}{plot panel background; options are "blank", "light", 
or "dark"}
}
\value{
a theme object
}
\description{
A minimal theme for most charts that uses PYN's fonts
}
