% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtc_colors.R
\name{color_jtc}
\alias{color_jtc}
\title{Add PYN colors to a chart}
\usage{
color_jtc(type = "fill", scale = "primary", rev = FALSE)
}
\arguments{
\item{type}{"fill" for use with the fill aesthetic; 
"color" for use with the color aesthetic,}

\item{scale}{"primary" (default), "dark", "light", "greys", "blues", "oranges",
"tans", "yellows"}

\item{rev}{reverse the order of the colors; defaults to false}
}
\value{
a manual scale object
}
\description{
scales for dark, light, and primary have 5 colors, color scales have three shades;
if more are needed, use scale_X_manual
}
