% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jtc.R
\name{theme_jtc}
\alias{theme_jtc}
\title{PYN Theme}
\usage{
theme_jtc(base_family = "franklin", base_size = 14,
  base_color = "#4f4f4f", axis.title.y = element_text(hjust = 1, vjust
  = 2, face = "bold"), axis.title.x = element_text(hjust = 1, face =
  "bold"), plot.title = element_text(hjust = 0, size = 20, face =
  "bold"), plot.subtitle = element_text(hjust = 0, size = 16),
  legend.position = "right", background = "blank")
}
\arguments{
\item{background}{plot panel background; options are "blank", "light", 
or "dark"}
}
\value{
a theme object
}
\description{
A minimal theme for most charts that uses Libre Franklin font
}
